+++ WebAPI/BookingYacht/BookingYacht.API/Controllers/Admin/VehicleController.cs
@@ -1,12 +1,12 @@
-﻿using BookingYacht.Business.Interfaces.Agency;
+﻿using System;
+using System.Threading.Tasks;
+using BookingYacht.Business.Interfaces.Admin;
 using BookingYacht.Business.SearchModels;
 using BookingYacht.Business.ViewModels;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
-using System;
-using System.Threading.Tasks;
 
-namespace BookingYacht.API.Controllers.Agency
+namespace BookingYacht.API.Controllers.Admin
 {
     [Route(ApiVer2Route)]
     [ApiController]
